asv_defence/requirements.txt
colorlog==5.0.1
hydra-colorlog==1.1.0
hydra-core==1.1.1
hyperlink==17.3.1
HyperPyYAML==1.0.0
numpy<2.0
opencv-python
soundfile
sox
torch==1.12.1+cu113
torchvision==0.13.1+cu113
torchaudio==0.12.1
tqdm
wandb
lib